Multiple-Disease-Prediction-Webapp/Frontend/requirements.txt
# Core Dependencies
streamlit==1.28.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
plotly==5.15.0
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Traditional ML Libraries
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2

# Deep Learning Dependencies
tensorflow-cpu==2.13.0
keras==2.13.1
opencv-python==4.8.0.76

# Image Processing
imageio==2.31.1
scikit-image==0.21.0

# Streamlit Extensions
streamlit-option-menu==0.3.6

# Utilities
tqdm==4.66.1
